# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nexus Bank AI API",
    "description": "Platform API for Nexus Bank AI services. KYC (face, OCR, liveness) is live. AML screening will be added as a separate domain.",
    "contact": {
      "name": "Nexus Bank"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/kyc/verify": {
      "post": {
        "tags": [
          "KYC"
        ],
        "summary": "Verify identity",
        "description": "Face detection + face match + OCR on an ID document and a selfie.",
        "operationId": "verify_kyc_api_v1_kyc_verify_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_verify_kyc_api_v1_kyc_verify_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KYCVerificationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kyc/ocr": {
      "post": {
        "tags": [
          "KYC"
        ],
        "summary": "Extract document fields",
        "operationId": "extract_ocr_api_v1_kyc_ocr_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_extract_ocr_api_v1_kyc_ocr_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OCROnlyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_extract_ocr_api_v1_kyc_ocr_post": {
        "properties": {
          "document": {
            "type": "string",
            "format": "binary",
            "title": "Document",
            "description": "Document image for OCR"
          }
        },
        "type": "object",
        "required": [
          "document"
        ],
        "title": "Body_extract_ocr_api_v1_kyc_ocr_post"
      },
...
      "Body_verify_kyc_api_v1_kyc_verify_post": {
        "properties": {
          "id_document": {
            "type": "string",
            "format": "binary",
            "title": "Id Document",
            "description": "ID card or passport image"
          },
          "selfie_image": {
            "type": "string",
            "format": "binary",
            "title": "Selfie Image",
            "description": "Selfie photo"
          }
        },
        "type": "object",
        "required": [
          "id_document",
          "selfie_image"
        ],
        "title": "Body_verify_kyc_api_v1_kyc_verify_post"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error message"
          },
          "details": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "example": {
          "details": {
            "confidence_threshold": 0.6
          },
          "error": "ValidationError",
          "message": "No face detected in ID document",
          "timestamp": "2024-10-11T10:30:00Z"
        }
      },
      "FaceMatchData": {
        "properties": {
          "verified": {
            "type": "boolean",
            "title": "Verified",
            "description": "Whether faces match"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Match confidence score"
          },
          "similarity_metrics": {
            "$ref": "#/components/schemas/SimilarityMetrics"
          },
          "threshold_used": {
            "type": "number",
            "title": "Threshold Used"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "verified",
          "confidence",
          "similarity_metrics",
          "threshold_used",
          "message"
        ],
        "title": "FaceMatchData"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KYCVerificationResponse": {
        "properties": {
          "verification_status": {
            "$ref": "#/components/schemas/VerificationStatus"
          },
          "confidence_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Score"
          },
          "face_match_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Face Match Score"
          },
          "ocr_data": {
            "$ref": "#/components/schemas/OCRData"
          },
          "processing_time_ms": {
            "type": "integer",
            "title": "Processing Time Ms"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "face_verification_details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FaceMatchData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "verification_status",
          "confidence_score",
          "face_match_score",
          "ocr_data",
          "processing_time_ms"
        ],
        "title": "KYCVerificationResponse"
      },
...
      "OCRData": {
        "properties": {
          "document_type": {
            "type": "string",
            "title": "Document Type",
            "description": "Detected document type"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Overall OCR confidence"
          },
          "extracted_text": {
            "type": "string",
            "title": "Extracted Text",
            "description": "Full raw text extracted"
          },
          "fields": {
            "allOf": [
              {
                "$ref": "#/components/schemas/OCRFields"
              }
            ],
            "description": "Structured extracted fields"
          }
        },
        "type": "object",
        "required": [
          "document_type",
          "confidence",
          "extracted_text",
          "fields"
        ],
        "title": "OCRData"
      },
      "OCRFields": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "document_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Number"
          },
          "nationality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nationality"
          },
          "issue_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Date"
          },
          "expiry_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Date"
          },
          "place_of_birth": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Of Birth"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          }
        },
        "type": "object",
        "title": "OCRFields"
      },
      "OCROnlyResponse": {
        "properties": {
          "ocr_data": {
            "$ref": "#/components/schemas/OCRData"
          },
          "processing_time_ms": {
            "type": "integer",
            "title": "Processing Time Ms"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "ocr_data",
          "processing_time_ms"
        ],
        "title": "OCROnlyResponse"
      },
      "SimilarityMetrics": {
        "properties": {
          "cosine_similarity": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Cosine Similarity"
          },
          "euclidean_distance": {
            "type": "number",
            "minimum": 0.0,
            "title": "Euclidean Distance"
          }
        },
        "type": "object",
        "required": [
          "cosine_similarity",
          "euclidean_distance"
        ],
        "title": "SimilarityMetrics"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerificationStatus": {
        "type": "string",
        "enum": [
          "approved",
          "rejected",
          "pending",
          "error"
        ],
        "title": "VerificationStatus"
      }
    }
  },
...
}